Enhance testing capabilities by adding agent trace fixture and snapshot tests
- Introduced a new `agent_trace` fixture in `conftest.py` to load trace data from JSON files for testing.
- Updated `test_trace.py` to utilize the new fixture, improving snapshot testing for agent traces.
- Removed the old `conftest.py` from the unit tests as it was no longer needed.
- Updated snapshot files to reflect changes in the agent trace structure and data.
